requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
dataclasses-json==0.6.4
diskcache==5.6.3
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.43
langchain-text-splitters==0.0.1
langsmith==0.1.48
llama-cpp-python==0.2.61
lxml==5.2.1
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.1
packaging==23.2
pdfminer.six==20231228
pycparser==2.22
pydantic==2.7.0
pydantic-core==2.18.1
pypdf==4.2.0
python-docx==1.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy==2.0.29
tenacity==8.2.3
typing-extensions==4.11.0
typing-inspect==0.9.0
urllib3==2.2.1
yarl==1.9.4
